Pass fixed obsevables to n3fit
We change somewhat the interface w.r.t the usual datasets, by making use
of the FixedObservableData object. One annoyance is that we have to
manually repartition the output of make_replica to get the right central
values. This should get fixed with a new implementation that returns the
fluctuated objects (when we get around having DatasetData).