Enable fixed observables in fit
The main complication is dealing with the keras model logic and the meta
wrapper along it.

We cannot just have a constant input for a layer but instead we need to
pass it as input to the model as well.

The model wrapper, "meta model", had some particular logic to deal with
this, but I find it too confusing, so mostly work around it.

To that end we pull the constant predictions out of the observable
wrapper and pass them in again where needed.